Refactored to include Router
