+++ frontend/package.json
@@ -11,7 +11,9 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^6.16.0"
   },
   "scripts": {
     "start": "react-scripts start",
