+++ frontend/src/components/Schedule/schedule.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+Function Schedule() {
+
+    const renderSchedule = async (e) => {
+        return (
+            <div>
+            <h1>Schedule Page</h1>
+            <p>This is the schedule page.</p>
+            </div>
+        )
+    }
+}
+
+export default Schedule
