--- frontend/src/schedule.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-const Schedule = () => {
-    return (
-        <div>
-        <h1>Schedule Page</h1>
-        <p>This is the schedule page.</p>
-        </div>
-    )
-}
-
-export default Schedule
